+++ src/resources/views/livewire/parts/mobile-cards.blade.php
@@ -9,8 +9,10 @@
 
             <div class="flex justify-between items-center mb-2 px-4 py-3 rounded-t-lg {{ $themeConfig['table']['tr_thead'] }}">
                 <div 
-                    class="font-bold text-lg break-words flex-1 pr-2 cursor-pointer flex items-center gap-2 group"
-                    wire:click="openMobileCardDetails('{{ $row[$column_id] }}')"
+                    class="font-bold text-lg break-words flex-1 pr-2 flex items-center gap-2 group {{ $titleColumn && $titleColumn->triggerCardInfoModal ? 'cursor-pointer' : '' }}"
+                    @if($titleColumn && $titleColumn->triggerCardInfoModal)
+                        wire:click="openMobileCardDetails('{{ $row[$column_id] }}')"
+                    @endif
                 >
                     <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5 opacity-50 group-hover:opacity-100 transition-opacity">
                       <path stroke-linecap="round" stroke-linejoin="round" d="M2.036 12.322a1.012 1.012 0 0 1 0-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178Z" />
@@ -18,7 +20,9 @@
                     </svg>
 
                     @if($titleColumn)
-                        @if(property_exists($titleColumn, 'hasView') && $titleColumn->hasView)
+                        @if(isset($row[$titleColumn->key . '_card_title']))
+                             {!! $row[$titleColumn->key . '_card_title'] !!}
+                        @elseif(property_exists($titleColumn, 'hasView') && $titleColumn->hasView)
                             @include($titleColumn->view)
                         @elseif(property_exists($titleColumn, 'isHtml') && $titleColumn->isHtml)
                             {!! $row[$titleColumn->key] ?? '' !!}